Add car relations handling and update edit functionality
- Introduced `GetCarWithRelationsByIdQueryHandler` for fetching cars with related data.
- Updated `ICarRepository` with a new method for retrieving car relations.
- Modified `ResultCarWithRelationsDto` to use `byte` for `seat` and `luggage`.
- Enhanced `UpdateCarDto` to include a list of car features.
- Updated `UpdateCarFeatureDto` to add a reference to `ResultFeatureDto`.
- Renamed `Update` method to `Edit` in `AdminCarController` and adjusted API calls.
- Implemented logic in `Edit` to handle car feature submissions.
- Added `GetCarWithRelationsById` method in `CarsController`.
- Created a new Razor view `Update.cshtml` for car editing.

+++ Frontends/CarBook_OnionArch.Dto/CarDtos/ResultCarWithRelationsDto.cs
@@ -2,13 +2,12 @@
 {
     public class ResultCarWithRelationsDto
     {
-
         public int id { get; set; }
         public string? model { get; set; }
         public int mileage { get; set; }
         public string? transmission { get; set; }
-        public int seat { get; set; }
-        public int luggage { get; set; }
+        public byte seat { get; set; }
+        public byte luggage { get; set; }
         public string? fuel { get; set; }
         public string? coverImageUrl { get; set; }
         public string? bigImageUrl { get; set; }
